build and debug

+++ silvia_arduino/silvia_temperature_sensor.cpp
@@ -30,7 +30,7 @@ double TemperatureSensor::updateTemperature() {
     /*
     Update temperature using time averaging and filter based upon last reading
     */
-    double average = reading_sum_ / reading_count_;
+    double average = reading_sum_ / (double)reading_count_;
     // Apply smoothing
     // ***************************************************** Inclusion here causes NAN
     reading_last_ = average * smoothing_filter_val_ + \
